[curry/trunk@1873]:
Ignore-this: d960b3a3fdde759b6a711b4d5c37dae9
Include definitions of ($!!), ($#), ($##), and unknown in the prelude,
which were added in version 0.8.2 of the Curry report, but not yet
part of the prelude. In addition, the primitive function ground is now
defined in the prelude. It is still exported from module Success for
backward compatibility.

* curry/Prelude.curry(ground, ($!!), ($#), ($##), unknown),
  curry/Success.curry(ground)

darcs-hash:20060316120105-4197f-b37e455bfbff53294750af55a4bc03ad2c108e53