[curry/trunk@1906]:
Ignore-this: 2bd07bdfd28afe12aee1634cc8e23ec6
Fix an issue where the shell scripts generated by the configure script
did not work when using autoconf version 2.52. This was due to using
the autoconf output variables, @abs_srcdir@, @abs_builddir@, and
@abs_top_builddir@, which are defined for autoconf version 2.57, but
not for version 2.52.

An additional benefit of this change is that the source directory can
be moved to a different place without invalidating the generated
scripts unless the configure script was invoked with an absolute path.

* configure.in, configure-bin.in, Makefile.in,
  cyc.in, cymake.in, cyi.in,
  in-place-config, in-place-config-bin:
  The in-place configuration is now prepared by including the
  appropriate settings at configuration time rather than sourcing them
  when the shell scripts are executed. The path to the build directory
  is now determined by examining the shell script variable $0.

darcs-hash:20060427092100-4197f-8835e3d18a927da6aeb7bfe669f6f091ada28830