[curry/trunk@2606]:
Ignore-this: e71707f8194e8b30ecaa5cf6e232245f
Setting the environment variable HC is now the preferred way to select
a particular Haskell compiler. The option --with-hc is deprecated and
will go away soon. For now it just prints a warning when it is
used. In addition, the code for the automatic detection of a suitable
Haskell 98 compiler has been streamlined a bit.

* configure.in,
  aclocal.m4(CURRY_PROG_GHC, CURRY_PROG_HBC, CURRY_PROG_NHC):
  These macros now set HC if their respective compiler supports
  Haskell 98 and they no longer determine the compiler version.
  Furthermore, CURRY_PROG_NHC now checks only for the command name
  nhc98.

* configure.in,
  aclocal.m4(CURRY_HC_VERSION, CURRY_GHC_IOEXTS):
  The Haskell 98 compiler command is now always bound to the shell
  variable HC.

* aclocal.m4(CURRY_HC_VERSION):
  The compiler version string is now cached.

darcs-hash:20080203215744-4197f-e7d3040bef2736b4b8b79cb41f1e8ea32001f141