[curry/trunk@2980]:
Ignore-this: 8325e72245230199443f1a82079fcc
Systematically rename local variables that are introduced by the
compiler. This fixes an issue where the compiler would fail with an
internal error when abstracting a local variable declaration that was
introduced by the simplifier and then eta-expanded. In addition, this
ensures that the eta-expanded function is not exported.

* Desugar.lhs(freshVar),
  Records.lhs(freshVar),
  Newtype.lhs(freshVar),
  LazyPatterns.lhs(freshVar),
  CaseMatch.lhs(freshVar),
  Simplify.lhs(freshVar),
  PatternBind.lhs(freshVar)

darcs-hash:20100709134537-4197f-9b3f74c7f75ec031c669395e130b7a0027c26548