[curry/trunk@3010]:
Ignore-this: 75a51776d2c2b0d195ace657cff24200
Represent a renaming type which is exported abstractly as an abstract
data type in the module interface. This improves interface stability
and restores the export policy before r2492.
An important reason for the change in r2492 was to pave way for using
renamed basic foreign types in foreign function declarations, which,
as specified in version 1.0 of the Haskell FFI addendum, was allowed
even for renaming types whose constructor is not exported. This has
been corrected in the Haskell 2010 FFI spec.

* IntfCheck.lhs(checkImport):
  Once again accept an abstract data type declaration as
  representation of a renaming type.

* Exports.lhs(typeDecl):
  Represent an abstract renaming type as abstract data type.

* TypeInfo.lhs:
  Update documentation.

darcs-hash:20101004095449-4197f-b69220104e285ec05e70b98fa4cedeed52b00b41