[curry/trunk@3200]:
Ignore-this: 76b00fd48e8e4bc7a7e57804481afb17
Fix matching of function patterns in case expressions. The system now
uses a rigid variant (==<=) of the function pattern constraint (=:<=)
to ensure that a variable in the matched expression is never
instantiated to a term from the matched pattern. In addition,
evaluation no longer needlessly suspends when the matched pattern
reduces to a free variable and the matched expression is a free
variable, too, as in the contrived goal case _ of (id x) -> x.

* CaseMatch.lhs(matchFlex, matchRigid, inject, unifyRigid),
  runtime/funpat.nw(flex_pair_cons, rigid_pair_cons, ___61__58__60__61_,
                    ___61__61__60__61_, funpat_match_1, funpat_match_2,
                    funpat_bindVars, funpat_bindVars_1, funpat_bindVarsRigid,
                    funpat_bindVarsRigid_1),
  curry/Prelude.curry((==<=)):
  Use the new primitive (==<=) to expand function patterns in case
  expressions instead of merely applying ensureNotFree to the matched
  expression.

* curry/Prelude.icurry-boot:
  Updated by the build process.

darcs-hash:20160409165420-4197f-1444da460d0d8239e93742efe1eb2378b60bf580